requirements.txt
# dependencies used in all environments

# Non-Python dependencies
# sudo apt-get install python2.7, supervisor, imagemagick

                                   # LICENSE  Python3
beautifulsoup4==4.6.0              # MIT      y
django>=3.0.0,<3.1                 # MIT      y
django-bootstrap-pagination==1.6.4 # MIT      ???
django-cors-headers==3.3.0         # MIT      y
django-redis==4.12.1               # BSD      y
django-redis-sessions==0.6.1       # BSD      y
djangorestframework>=3.11.0,<3.12  # BSD      y
drf-yasg>=1.17.0,<1.18             # BSD      y
elasticsearch>=7.0.0,<8.0.0     # Apache y        y   NOTE: match namesdb version
elasticsearch-dsl>=7.0.0,<8.0.0 # Apache y        y   NOTE: match namesdb version
envoy==0.0.3                       # MIT      ???
GitPython==3.1.1                   # BSD      y
gunicorn==19.8.1                   # MIT      y
redis>=3.5.0,<3.6                  # MIT      y
requests==2.22.0                   # Apache2  y
sorl-thumbnail==12.6.3             # BSD      y

bpython==0.17.1

coverage
pytest
pytest-cov
pytest-django
pytest-testmon
pytest-xdist
